pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "canvas-cli"
version = "0.1.0"
description = "MCP server exposing Canvas LMS data as tools for Tomo agent workflows"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Canvas CLI Team" }
]
dependencies = [
    "mcp>=1.0.0",
    "canvasapi==3.4.0",
    "pydantic>=2.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
canvas-cli = "canvas_cli.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/canvas_cli"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
